cli refactor
